Signout logic and buttons
Implemented logic to be able to sign out from the current supabase session in the app via a button in the Admin Center and the dashboard. For this, the login method needed to be updated because the supabase session that was created at login had previously not been held correctly, thus not making it possible to sign out, since there had not been a valid session.